# pulse-core-nextjs/openapi.yaml
openapi: 3.0.3
info:
  title: AfyaHero FHIR API
  description: |
    Healthcare platform API with FHIR interoperability.
    
    ## Authentication
    All endpoints require a valid Bearer token obtained through Supabase authentication.
    
    ## Rate Limiting
    API calls are rate limited based on user role and endpoint type.
  version: 1.0.0
  contact:
    name: AfyaHero Support
    email: support@afyahero.com
servers:
  - url: https://api.afyahero.com
    description: Production server
  - url: https://staging.api.afyahero.com
    description: Staging server
  - url: http://localhost:3000
    description: Local development
...
paths:
...
  /api/apps/auth/introspect:
    post:
      tags:
        - Apps Auth
      summary: Introspect app token
      description: Returns status and claims for an app bearer token.
      security:
        - AppBearerAuth: []
      responses:
        '200':
          description: Token introspection result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppTokenIntrospectionResponse'
        '400':
          $ref: '#/components/responses/AppBadRequest'
        '401':
          $ref: '#/components/responses/AppUnauthorized'
        '403':
          $ref: '#/components/responses/AppForbidden'
        '429':
          $ref: '#/components/responses/AppRateLimited'
...
  /api/apps/chp/visit-notes:
    get:
      tags:
        - Apps CHP
      summary: List CHP visit notes
      description: Returns visit notes for CHP workflows.
      security:
        - AppBearerAuth: []
      responses:
        '200':
          description: Visit note list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppVisitNotesResponse'
        '400':
          $ref: '#/components/responses/AppBadRequest'
        '401':
          $ref: '#/components/responses/AppUnauthorized'
        '403':
          $ref: '#/components/responses/AppForbidden'
        '429':
          $ref: '#/components/responses/AppRateLimited'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Supabase JWT token
    DeveloperSignedHeaderAuth:
      type: apiKey
      in: header
      name: X-Developer-Signature
      description: Developer signature header used for app token issuance.
    AppBearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: App access token for /api/apps endpoints.
  responses:
    AppBadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/AppErrorResponse'
    AppUnauthorized:
      description: Authentication failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/AppErrorResponse'
    AppForbidden:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/AppErrorResponse'
    AppRateLimited:
      description: Too many requests
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/AppErrorResponse'
  schemas:
    AppErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        code:
          type: string
      required:
        - error
        - message
...
    AppTokenIntrospectionResponse:
      type: object
      properties:
        active:
          type: boolean
        subject:
          type: string
        scope:
          type: string
        exp:
          type: integer
      required:
        - active
...
    AppVisitNotesResponse:
      type: object
      properties:
        notes:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
              patientId:
                type: string
              createdAt:
                type: string
                format: date-time
              note:
                type: string
            required:
              - id
              - patientId
              - createdAt
      required:
        - notes
...
security:
  - BearerAuth: []